python_card_parser/requirements.txt
antlr4-tools==0.2.1
antlr4-python3-runtime==4.13.0
